refactor: 인증 쿠키 문자열 생성 로직 유틸 함수로 분리
- accessToken, refreshToken 쿠키를 문자열로 반환하는 getAuthCookieHeader 함수 생성
- 서버 컴포넌트에서 인증 헤더 생성 시 활용 가능
